--- @types/remark-html.d.ts
@@ -1,1 +1,0 @@
-declare module 'remark-html'
