--- components/typography/Text/text.module.scss
@@ -1,71 +1,0 @@
-.h1 {
-  font-size: 6.25rem;
-  padding-right: 2rem;
-  letter-spacing: -0.04em;
-  line-height: 1.2;
-  font-weight: 700;
-}
-
-.h2 {
-  font-size: 4.5rem;
-  letter-spacing: -0.04em;
-  line-height: 1.2;
-  font-weight: 700;
-}
-
-.h3 {
-  letter-spacing: -0.04em;
-  font-size: 2.25rem;
-  line-height: 2.5rem;
-  font-weight: 700;
-}
-
-.h4 {
-  font-size: 2.125rem;
-  font-weight: 500;
-}
-
-.h5 {
-  font-size: 1.5rem;
-  font-weight: 500;
-}
-
-.h6 {
-  font-size: 1.25rem;
-  font-weight: 500;
-}
-
-.subheading1 {
-  font-size: 1rem;
-  font-weight: 500;
-}
-
-.subheading2 {
-  font-size: 0.875rem;
-  font-weight: 500;
-}
-
-.body1 {
-  font-size: 1rem;
-}
-
-.body1 {
-  font-size: 0.875rem;
-}
-
-@media screen and (max-width: 1024px) {
-  .h1 {
-    text-align: center;
-    font-size: 4.5rem;
-    font-weight: 700;
-    line-height: 1.2;
-    letter-spacing: -0.04em;
-  }
-
-  .h2 {
-    letter-spacing: -0.04em;
-    line-height: 1.2;
-    font-weight: 700;
-    font-size: 2.5rem;
-  }
-}
