--- interfaces/note.ts
@@ -1,10 +1,0 @@
-
-type NoteType = {
-  slug: string
-  title: string
-  excerpt: string
-  content: string
-  source: string
-}
-
-export default NoteType
