--- interfaces/post.ts
@@ -1,14 +1,0 @@
-type PostType = {
-  slug: string
-  title: string
-  date: string
-  coverImage: string
-  excerpt: string
-  ogImage: {
-    url: string
-  }
-  content: string
-}
-
-export default PostType
-
