+++ src/components/templates/layout.tsx
@@ -1,0 +1,19 @@
+import Meta from '../meta'
+
+type Props = {
+  preview?: boolean
+  children: React.ReactNode
+}
+
+const Layout = ({ preview, children }: Props) => {
+  return (
+    <>
+      <Meta />
+      <div className="min-h-screen">
+        <main>{children}</main>
+      </div>
+    </>
+  )
+}
+
+export default Layout
