+++ src/helpers/markdownToHtml.ts
@@ -1,0 +1,7 @@
+import { remark } from 'remark'
+import html from 'remark-html'
+
+export default async function markdownToHtml(markdown: string) {
+  const result = await remark().use(html).process(markdown)
+  return result.toString()
+}
