+++ src/interfaces/note.ts
@@ -1,0 +1,10 @@
+
+type NoteType = {
+  slug: string
+  title: string
+  excerpt: string
+  content: string
+  source: string
+}
+
+export default NoteType
