+++ src/interfaces/presentation.ts
@@ -1,0 +1,14 @@
+type PresentationType = {
+    slug: string
+    title: string
+    date: string
+    coverImage: string
+    excerpt: string
+    ogImage: {
+      url: string
+    }
+    content: string
+  }
+  
+  export default PresentationType
+
