+++ src/modules/home.tsx
@@ -1,0 +1,4 @@
+// TODO - put info about me
+export default function Home() {
+  return null;
+}
