+++ src/services/api.ts
@@ -1,0 +1,49 @@
+import fs from "fs";
+import { join } from "path";
+import matter from "gray-matter";
+
+// Get Directory Path
+const presentationsDirectory = join(process.cwd(), "_presentations");
+
+// Read Directory Slugs
+export function getPresentationSlugs() {
+  return fs.readdirSync(presentationsDirectory);
+}
+
+// Read MD Content
+export function getPresentationBySlugs(slug: string, fields: string[] = []) {
+  const realSlug = slug.replace(/\.md$/, "");
+  const fullPath = join(presentationsDirectory, `${realSlug}.md`);
+  const fileContents = fs.readFileSync(fullPath, "utf8");
+  const { data, content } = matter(fileContents);
+
+  type Items = {
+    [key: string]: string;
+  };
+
+  const items: Items = {};
+
+  // Ensure only the minimal needed data is exposed
+  fields.forEach((field) => {
+    if (field === "slug") {
+      items[field] = realSlug;
+    }
+    if (field === "content") {
+      items[field] = content;
+    }
+
+    if (typeof data[field] !== "undefined") {
+      items[field] = data[field];
+    }
+  });
+
+  return items;
+}
+
+export function getAllPresentations(fields: string[] = []) {
+  const slugs = getPresentationSlugs();
+  const presentations = slugs.map((slug) =>
+    getPresentationBySlugs(slug, fields)
+  );
+  return presentations;
+}
