--- styles/globals.scss
@@ -1,19 +1,0 @@
-// COLOR ------------------------------------------------------------------------------------------------ 
-$color-black: #000;
-$color-gray: #b6b6b6;
-$color-white: #fff;
-
-// TYPOGRAPHY ------------------------------------------------------------------------------------------- 
-// style
-
-// size
-
-// weight
-
-// BREAKPOINTS ------------------------------------------------------------------------------------------
-// mobile
-$bp-sm: 576px;
-// tablet
-$bp-md: 768px;
-// desktop [992px+]
-$bp-lg: 992px;
