--- styles/markdown-styles.module.css
@@ -1,42 +1,0 @@
-.markdown {
-    @apply text-lg leading-relaxed;
-  }
-
-.markdown a {
-  color: grey;
-}
-  
-.markdown p,
-.markdown img {
-  max-width: 100%;
-  height: auto;
-  padding: 5px 0 5px 0;
-}
-.markdown ul {
-  list-stylse: auto;
-  margin: auto;
-  padding: revert;
-}
-
-.markdown ol {
-  list-style: auto;
-  margin: auto;
-  padding: revert;
-}
-
-.markdown blockquote {
-  @apply my-6;
-}
-
-.markdown h1 {
-  @apply text-4xl font-bold mt-12 mb-4 leading-snug;
-}
-
-
-.markdown h2 {
-  @apply text-3xl mt-8 mb-4 leading-snug;
-}
-
-.markdown h3 {
-  @apply text-2xl mt-8 mb-4 leading-snug;
-}
